Revert pf9-hostagent to using PF9_VERSION (for now)
We realized that HA contains certificates, and the HA package version
does not reflect certificate content. So the only way to ensure that
certificates are upgraded if they've changed is to ensure HA itself
is upgraded each release. To be safe, go back to tying the HA version
to the product version to force the upgrade, until we find a way
to separate certs from HA.