Handle potential NULL values in resmgr's JSON fields in DB
Handle potential NULL values in resmgr's JSON fields in DB. Some of the JSON
role config data that is stored by resmgr are read from and written to the DB by
the JsonBlob class. When reading out data which could be NULL, return None
instead of attempting to do a json.loads with it. In case of writes, pass in
None instead of trying to do a json.dumps with it.

This was noticed in a DU where some role rabbit_permissions and custom_settings
were set to NULL and when resmgr tries to read all role data to determine latest
version of a role, it would throw a json.loads exception and crash.

Issues addressed: IAAS-4683

Testing done: Patched a failed DU where reading in all the roles data (to
determine latest version of role) was failing. Ensured it passed after this
change. No easy way to test the write flow part of this change. Passed private
build:
http://teamcity.platform9.sys:8111/viewLog.html?tab=buildLog&buildTypeId=Pf9project_DuAmi&buildId=212533

Reviewed by: leb, rdeuel, osoriano (https://rbcommons.com/s/platform9/r/3396/)