IAAS-4801 Host agent extension for iface:ipaddress list
A new host agent extension to retrieve the interface name along
with the ip address. This is similar to the fetch_ip_address.py
with the difference that this also sends out bridges that are
created using ovs whether they have an ip or this.

This is part of user experience improvement for Neutron.
over time, fetch_ip_address.py can be retired.

Testing done:
Manually added the script in a live hypervisor and confirmed that
the agent returns the required values.
Also tried on a hypervisor without ovs to make sure the ip address
with interface name list is returned with an empty list for ovs_bridges.