Reset converge attempts when session.start is called
session.start runs in a loop, but the number of converged attempts
is not reset every time it is called. This can lead to hostagent
being stuck in the 'failed' state and not converging.

Hostagent should try to converge since starting comms could fix the
host communication.

Testing done:
- Reproduced the issue
- Made the change and verified hostagent would still restart comms

Issues: IAAS-5054