Fixing error check with process communicate
process.communicate() returns the stdout, stderr tuples. When a
command generates stderr it is returned by process.communicate. The
command might have succeeded. process.returncode specifies if the
command executed without an error.

Change: Use process.returncode instead of stderr.

Earlier my code did process the output if there was anything in stderr
but this seems to cause issues when sudo prints error strings.

Cause for change:
When hostname is changed on a host and the DNS entry for the same
(127.0.0.1) isn't fixed as well, sudo will throw:
'sudo: unable to resolve host ubuntu08\n'
error. This halted ovs-bridge collection.