Remove assumption that app name starts with pf9 during erase.
The pf9-yum and pf9-apt scripts internally assume that the packages they are
trying to erase start with pf9. This assumption causes issues when we integrate
with partner apps. Current such case with midonet is leading to deauth
operations failing because their apps are called midolman and mem_repo. This
change removes such a restriction in those scripts.

This should not open a security hole since the sudoers file packaged with
hostagent restricts these erase commands to packages starting with pf9. For
midonet apps, we will have the app provide its own sudoers file that will allow
erasure of its own packages.

Testing done: Tried this out with the midonet's apps

Reviewed by: osoriano, leb (https://rbcommons.com/s/platform9/r/4142/)