Disable janitor operations on Neutron setup.
This is a temporary workaround for the IAAS-5438 issue. Janitor cleanup of
instances can trigger a side effect where the networking interfaces to VMs
on the host would be deleted.

This change makes Janitor do nothing in case Neutron is setup on the DU.

Issues addressed: IAAS-5438

Testing done: Ensured Janitor continues its loop on a non Neutron DU and it
doesn't do anything in a Neutron DU.

Reviewed by: sarun, akapoor, pooja (https://rbcommons.com/s/platform9/r/4159/)