Fix janitor check for Neutron
Turns out in some initial boot time cases systemd may report a service as active
even though it is not running or enabled. To avoid this race condition, use
crudini to check the nova.conf setting to see if it is using neutron network
driver.

Testing done: Patched a DU and ensured it detects the Neutron setup well.

Issues addressed: IAAS-5441