IAAS-5727 : fix severe performance problem in hostagent
We discovered that on Ubuntu hosts with pf9-kube role, hostagent
takes around 25 (!) seconds to process a periodic heartbeat. During
this time, AMQP messages are not handled. This causes delays in
the processing of support commands, which qbert relies heavily on.
This causes timeouts and random integration test failures.
In general, this can also break other operations that rely on
support commands, such as remote ssh and bundle generation.

There are 2 root causes:

1. Hostagent queries the app and service state of the system
   TWICE per heartbeat.
   The first time is by get_current_config(), which is part
   of the main session.py code.
   The second time is a side effect of calling process_apps(),
   which is used to see if system is currently converged,
   and again to converge the system if it isn't already.
   process_apps() queries the installed pf9 apps and their
   service state.

2. Specifically to Ubuntu, AptPkgMgr.query_pf9_apps() is very
   slow because it iterates over the entire list of packages
   in the local APT package database. With some instrumentation,
   I found that there are typically more than 80,000 entries
   in the DB! Iterating over these to find which ones provide
   'pf9app' capability can take 5-8 seconds.
   Problem (1) mentioned above exacerbates this, since it has
   to be done twice.

Proposed solution:

1. Modify process_apps() to accept a 'current' app/service
   configuration to compare against, if available.
   Since the code expects the app entries of the current config to
   be objects of the "App" base class, I had to add a 'make_app'
   method to AppDb to convert a dictionary-based app
   configuration back to an object of type App.
   Another complication is the fact that some apps use
   the old 'running' service state (which assumes the app only
   has one service with the same name as the app), while newer ones
   use 'service states' (allowing multiple services per app),
   so more conversion code was added to handle this.

2. On Ubuntu, take advantage of Apt.cache.get_providing_packages('pf9app')
   to filter packages up-front to only those providing the 'pf9app'
   capability. This is almost instantaneous.

Results:

On a particular pf9-kube host that used to take 25 seconds to complete a
heartbeat, the combined optimizations reduced it to about 5 seconds.
A large part of this is to simply run 'service pf9-kube status'
which takes about 4-5 seconds.
With no pf9 apps installed, an 'empty' heartbeat was reduced
from 7-8 seconds to about 1-2 seconds.