Add warning messages for time drift
Hosts will now have a message dictionary which maps message levels (e.g.
warn, error) to list of messages.
Only the time drift message has been added.

Example messages for a host:

{
    'warn': ['The host clock is out of sync']
}

This message will appear whenever when the host is responding to the DU
in time,
but the host timestamp is too far/behind the time on the DU.

Once the host timestamp meets the right criteria, the message will
disappear.

Note that this changes the meaning of a "responding" host. A host will
now
also be responding if it is responding within the threshold according to
the
timestamp taken when the host message is recieved on the DU.

Also, remove the host status field since it is redundant (equal to len
of roles)

Issues: IAAS-4590

Testing done:
  Passed du-ami build
  Tested manually using the UI