Fix janitor to not remove nova service in NSX deployment
Janitor currently looks only for the ostackhost-vmw role but in
case of the NSX deployment the role is ostackhost-neutron-vmw.
Hence janitor removes the authorized nova compute service from the
nova services table as part of clean up. With this fix janitor now
works with either of the roles being present on a host.

Bugs addressed:
IAAS-5502

Testing:
1. Deployed the code on a NSX DU and verified that the nova compute
record was not removed from nova database.

Reviewed by:
Sachin and Susmitha
https://rbcommons.com/s/platform9/r/4332/