Fix janitor logging
fixes https://platform9.atlassian.net/browse/IAAS-5627

The main problem was that we were reading the config as
a string and then multiplying it by 1024. This resulted in garbage
that effectively disabled rotation.

Also cleaned things up and put the logs in their own directory.

testing: built the installer and verified the new config was installed
properly. Set the maxKBytes value to 1 and validated rotation.
Introduced an error in the config and verified it showed up in
janitor-out.log