Add role states to the resmgr db schema
The original goal of this change was to add an atrribute to the
host_role_map table to track the application/update/removal of roles
from a host. Unfortunately, this changes the way sqlalchemy binds
together hosts and roles in the object mapping, so it ended up being
bigger than I'd have hoped.

Because of the extent of the changes, I also added unit tests
for a big part of the dbutils module (around 75% coverage). This
required that I make the code work with sqlite, so I had to patch
up some of the code that used mysql-specific statements.

testing: unit tests, hand testing on a DU with role add, upgrade and
remove. Waiting for a du-ami build to complete.