Implement role state machine in resmgr
This change adds the state management to resource manager
role application. It holds a state variable in the database
associated with the host/role association, and moves from
state to state only where it makes sense. This restricts/enables
behavior as follows:
+ When a role is in the process of converging as part of an add,
  edit, update or delete operation, no request will be accepted
  on the role.
+ When a request is made to add a role, an 'on_auth' event handler
  is run. If it fails, the request fails. Similar for delete with
  an on_death handler.
+ after a host converges, if a role was in the 'converging' state,
  an on_auth_converged or on_deauth_converged handler runs. If
  these fail, we enter an error state state that can only be
  recovered with a new add/delete operation.

In addition:
+ lots of changes to tests, cleaning up some boilerplate. More needed
+ modified the state machine so exiting an error state still runs the
   on_auth, on_deauth events. Updated here:
   https://platform9.atlassian.net/wiki/pages/viewpage.action?pageId=91258896
+ moved the event execution code the ResMgrProvider to the RolesMgr class
   so it can be used by the BbonePoller(process_existing_hosts) to run
   post-converge events.

TODO: If there's a crash, the state value for a role may be left in
a middle state. Need to implement _recover_unexpected_role_states
to move to a recoverable state in the bbone process_hosts loop.

testing: many new unit tests, private du-ami build