Fix pf9-cindervolume upgrade script part 2
To upgrade a role, the upgrade_hosts.py script loops over a host's
list of currently authorized roles and sends a put request for the same
rolename. This action triggers a resmgr update to the role.

Since the new roles do not exist yet, the upgrade script
won't be able to do so for our newly renamed and added roles.

The workaround is to add dummy roles with the equivalent rolenames.

Testing:
Passed an upgrade with VMware. The appliance initially had the
pf9-cindervolume role and after upgrade had both pf9-cindervolume-base
and pf9-cindervolume-lvm (since it was auth with the lvm driver).

OK with KVM (resmgr returned the correct roles and DB is as expected).
However, actual upgrade failed due to some sidekick issue.