Fix role deletion in resmgr
In delete_role and _process_exisiting_hosts, we were using
dbutils.query_host_and_app_details both to generate the host config
sent to bbmaster, as well as the config used for the deauth events.
In some cases, we want this to respect the role state, for example,
we don't want to push an app that's part of a role that's in the process
of being removed. OTOH when we're calling on_deauth_converged, we
need app configs for roles that are being removed.

In the spirit of trying to keep as much of the existing code as possible,
I've added a flag to change the behavior of query_host_and_app_details
under certain conditions. At some point in the future, this stuff
should be rethought from scratch.

testing: tweaked and ran unit tests, du-ami pending