Crash-proof resmgr role state machine
Since resource manager role application is now locked into a
state machine, for which only certain states are actionable from
the API, it's important that a crash in one of the non-terminal states
doesn't leave the machine stuck. This change adds a handler to
_process_existing_hosts that can handle all the 'middle' states
that might be left if an unexpected error/crash occurs.

This required that I factor out the significant code from the state
transitions so it could be used to advance the state, thus all the new
'move_to_...' functions.

Also added a pantload of unit tests.

testing: unit tests. integration build pending