Don't delete rabbit credentials on failed conflicting role removal
fix for: https://platform9.atlassian.net/browse/IAAS-6519

When a deauth fails because there's an auth in progress, resmgr
removes the rabbit credentials before failing. In this case, the
role is broken - it can't login to the rabbit broker. In the case
of nova, if this happens at the right time, nova-compute won't be
able to create a compute services in the database, and this breaks
the deauth hooks that may later attempt to force-down the service.
All in all we're left in a bad state that requires manual db
manipulation to get out of.

The change moves the updates to rabbit after the state check, so if
a role conflict is thrown, rabbit creds are safe.

testing: Added checks and verified in the unit tests