Fix installer unit tests for unresolvable FQDN
We were using squid.company.com but turns out that this is resolvable of late.
This is leading to the installer unit tests to fail. Changed it to an FQDN under
platform9.com domain, which we own and so that this doesn't happen in future.

Issues addressed: IAAS-6929

Testing done: Ran the unit test. Was failing before but succeeds now.

Reviewed by: vteves (https://rbcommons.com/s/platform9/r/5770/)