Ignore missing auth_events implementation file
fix for https://platform9.atlassian.net/browse/IAAS-7049

Auth events modules are version specific. Because of this, when we
try to remove a role from a previous version of the product, for
example if the host was down during upgrade, the events module
from the old version is gone and deauth fails. This ignores that
failure. This will be fairly rare, so I decided not to make a big
change to, say keep old events modules in the database (so they're
preserved across upgrades), or make the events files non-version
specific.

testing: Added a new unit test. Integration build pending.