Fix pf9 user creation in hostagent post install step for CentOS
It is ok if the home dir creation fails during pf9 user creation. It happens
when SELinux is on. We have steps after this to create the home dir explicitly.
The change keeps the same behavior as previous releases but will also work if we
set the -e flag in the post install script.

Testing done: Private build (failed test is not related to this change)
http://teamcity.platform9.sys:8111/viewLog.html?buildId=405129&buildTypeId=Pf9project_Integration_DuAmiIncr&tab=buildResultsDiv