Use a different hostagent restart method
The delayed restart we introduced in 3.0 seems to have many side effects
1) It assumes that 5 seconds is enough to commit the yum transaction. IAAS-8474
2) During these 5 seconds, hostagent may start downloading some package and that
can lead to a broken package download. IAAS-8378

The reliable way to perform this restart is to have this happen in a
synchronized fashion, as a last step in the hostagent upgrade logic. Given the
nature of hostagent upgrde, this code change will only be effective for upgrades
post this release. As a result, we need to keep the delayed restart logic for
one more release.

Testing done: Ran this hostagent restart logic in a loop (by manually triggering
it as part of the heartbeat operation). Verified that the restart logic did the
right thing. Also ran the upgrade build:
http://teamcity.platform9.sys:8111/viewLog.html?buildId=485863&tab=buildResultsDiv&buildTypeId=Pf9project_IntegrationBuildPipeline_016upgradeTests
(failures are not caused by my changes)

Some thoughts are captured here:
https://platform9.atlassian.net/wiki/spaces/~pshanbhag/pages/121904850/Migrating+host+side+components+to+systemd

Issues addressed: IAAS-8474, IAAS-8378.

Review: https://rbcommons.com/s/platform9/r/7907/