Change the hostagent restart method
With the move to systemd, some issues surfaced with hostagent restarts after
upgrading hostagent. The restart was previously done as the last step in the
hostagent rpm. With the systemd restart, this was leading to the yum transaction
getting aborted before it was committed to the yum db.

The change moves the restart out of the post install steps. It is part of the
hostagent code itself.

Also updated the sudoers file to add corresponding systemctl commands.

Testing done: Manual testing. Also ensure the discover tests run successfully on
both CentOS 7 and Ubuntu 14.

Issues addressed: IAAS-7782

Review: https://rbcommons.com/s/platform9/r/6753/