Add restart logic to hostagent post install
In 3.0, we moved the logic of hostagent restart from the post install script to
hostagent itself. However, this is useful only for upgrade from 3.0 onwards.
This is a temporary hostagent restart logic in the post install script for this
release. There is also an additional restart of the hostagent in the upgrade_hosts
script as a fallback in case the post install restart doesn't do the job.

Testing done: Ran a manual upgrade of the DU and ensured the hosts are upgraded
fine.

Issues: IAAS-7782

Review: https://rbcommons.com/s/platform9/r/6802/