Add `--force` option to `hostrole apply` command
When a pf9-kube role fails to converge and the role moves to the
`auth-error` state in resmgr, the only way to move it back to the
`start-apply` state is to apply the role.

Ordinarily, `du_ctl` does not apply the role if it is already
up-to-date. This adds the option to apply the role unconditionally.

Prior to this change, the alternative was a curl command:
```
curl -X PUT http://localhost:8083/v1/hosts/<id>/roles/pf9-kube -H "X-Auth-Token: <token>"
```

Also:
- Fix string formatting in print statements (switch from log to print
  calls broke string formatting)
- Remove duplicate comment