Add a way to build a cred-free hostagent rpm and shell script installer
details: https://platform9.atlassian.net/wiki/display/~rdeuel/Resource+Manager+in+a+Container#ResourceManagerinaContainer-Proposal

This change adds two new build targets in pf9-du in preparation for creating
a resource manager container:

In bbslave, modify the makefile and 45-customize-hostagent-rpm, so that
when HOSTAGENT_SKIP_CERTS is set, we get a hostagent without certs, and
without a customized hostagent.conf. Except for the conditionals, the logic
and hostagent.spec file are the same as what we use in the pf9-configure
ansible role that builds hostagent*.rpm. This gives us a hostagent package
with no credentials that can be used with any DU of the appropriate version.

In uber-rpm, add a build for platform9-installer-redhat.sh. I'm deferring
debian for now since we don't need it for CSPi atm. It uses the same logic
we use to build the package with ansible in the pf9-configure role. When built
like this (without the hostagent and comms rpms) it requires the user to provide
DU credentials, and it uses these to download the packages, along with the
certificates and hostagent.conf with rabbit credentials. The script installs the
uncustomized hostagent and comms packages, and then puts hostagent.conf and certs
in place.

The script requires a 'nocert-packagelist.rpm' text file in /opt/pf9/www/private
on the resmgr container. This will be put in place in the container build. Here's
a sample of the package list:

pf9-comms-2.0.0-444.5557e57.x86_64.rpm
pf9-hostagent-2.4.0-5000.e71c9e5.x86_64.rpm
etc/pf9/certs/hostagent/cert.pem
etc/pf9/certs/hostagent/key.pem
etc/pf9/certs/ca/cert.pem
etc/pf9/hostagent.conf

Note that these installers can't coexist with our current DU-style builds. A DU
won't get these build artifacts, and a container will never run the pf9-configure
customization role. It makes me itch a little that both types of hostagent installer
have the same package name, but the end result for a host using our standard
installer or this one should be the same, so I don't think it's a problem.

testing: provisioned a host with the new script, and verified the host was recognized
and could make rabbit and mysql connections.  Also ran a smoke build to make sure I didn't
break the current builds.