Resmgr shouldn't if role is incomplete
If the config parameters for a role are not available, resmgr should just skip
this role after logging it. Resmgr is supposed to do a best effort method on
loading all the roles that are complete.

Also, adjusted the tests to ensure to validate this behavior.

Issues: IAAS-8126

Testing done: Unit tests

Review: https://rbcommons.com/s/platform9/r/7129/