Add systemctl enable for hostagent even in the upgrade path
We moved from init scripts to systemctl. This changes from "chkconfig on" to
"systemctl enable" to ensure the service starts on boot. During this update of
the rpm, we moved the service to systemd but missed out enabling the service on
boot. This was being done only during the install of the hostagent.

Issues: IAAS-8110