Resmgr docker container skeleton
This change includes the initial build and configuration for a container
running resmgr/bbmaster. It includes a Dockerfile, a makefile to pull it
all together, and confd configuration to render config files and certs
onto the container filesystem. It depends on a schema for the consul
store described here:
https://platform9.atlassian.net/wiki/spaces/~rdeuel/pages/120041746/Running+Platform9+Services+in+Containers+-+Architecture+Summary#RunningPlatform9ServicesinContainers-ArchitectureSummary-AfterKeystoneinit-region

The build requires someone to provide the comms installer from outside,
either in an artifacts directory (a la teamcity) or through an environment
variable (COMMS_RPM).

todo:
+ supervisord config to get the services and confd started
+ init-region to create the database and request rabbit and keystone users
+ list-services

testing:
+ verified everything was in the right place
+ rendered all the configs/certs from a local consul store (using -onetime)