Fix host upgrades for 3.x
Change addresses multiple areas listed below

1) Decide if service or systemctl command should be used in hostagent

Hostagent has been using service commands to manage services of the apps.
However, with the some of the services moving to systemd, this is not being very
reliable. This change moves the logic to see if there is an init script present.
If present, it will call the service command. If not, it will invoke the
systemctl command.

2) Restart comms the right way in the hostagent post install script
Hostagent install restarts comms in case certs (delivered by hostagent) have
changed. However, intermixing systemctl commands with an init service leads to
unpredictable behavior and disaster. Changed the logic in the post install
script for hostagent to call service or systemctl depending on what type of
service is present.

3) Changed the logic for hostagent restart in the hostagent post install
In tests, we noticed that the hostagent restart was a bit unreliable. We were
ending up with multiple hostagent processes (like 1 monitored under init and
another under systemd). Instead of just a restart command, we now have a pkill
to terminate all existing hostagents and then perform a restart.

4) Couple of tweaks to the systemd unit files to invoke the service through bash
to get the redirection working for the daemon logs.

5) Added a couple of systemd commands to the support bundle output.

6) Added dependency of procps (Ubuntu) and procps-ng (CentOS) because we rely on
pgreps and pkills now.

Testing done:
https://mongo-dev.platform9.horse/logs/install/4e7bc303-6400-45f4-95d8-37f83f21899f.log

Issues: IAAS-8156, IAAS-8189

Review: https://rbcommons.com/s/platform9/r/7170/