Add kube wrapper to resmgr container
Although the actual kube host role package is served from privatefiles
in nginx, resmgr needs the json config from this package so it can
authorize the role.

Also a bit of cleanup and some clarifying comments.

testing: started the container and verified that the qbert role was loaded
into the resmgr roles table.