Add more network information for the host agent extension
Add more network information for the host agent extension

The current network information passed by the host agent extension is not enough
to determine which interface is service which network. This becomes relevant
when dealing with a host that is multi-homed.

This change ensures that the network mask and MAC is also returned, in addition
to the IPv4 address. The previous data is also retained currently, to maintain
backward compatibility.

Testing done: Ran the extension and validated the output.
{
    "iface_info": {
        "docker0": {
            "ifaces": [{
                "addr": "172.17.0.1",
                "broadcast": "172.17.0.1",
                "netmask": "255.255.0.0"
            }],
            "mac": "02:42:61:58:a8:25"
        },
        "ens192": {
            "ifaces": [{
                "addr": "10.1.10.232",
                "broadcast": "10.1.255.255",
                "netmask": "255.255.0.0"
            }],
            "mac": "00:50:56:a9:49:82"
        }
    },
    "iface_ip": {
        "docker0": "172.17.0.1",
        "ens192": "10.1.10.232"
    },
    "ovs_bridges": []
}

Issues: PAL-133