Fix hostagent uninstall on Ubuntu16
When we uninstall hostagent on Ubuntu16, first disable the service to not start
on boot and then remove the service files.

Before this change, things were done the other way around and that would fail on
Ubuntu16 and its systemd world of things. This was done correctly for CentOS
systems but was missed out for Ubuntu16

Testing done: Fixed the prerm script on the host and we were able to uninstall
the package.

Issues: IAAS-8320

Review: https://rbcommons.com/s/platform9/r/7498/