Allow deauth of a host that bbmaster has forgotten
fix for https://platform9.atlassian.net/browse/IAAS-8048

If a host is brought down, and then bbmaster is restarted, bbmaster's
in-memory state doesn't have any info on the host, and raises a
KeyError when an attempt is made to configure the host. An attempt to
deauth the roles or remove the host leaves the host's roles in the
pre-deauth state with no way out but to edit the database.

This change causes bbmaster to assume that a host exists, and create a
'missing' entry for it when it can't find a hostagent-reported host.
I had to make a few tweaks to allow resmgr to deal with the incomplete
host document provided by bbmaster in this case.

testing - to repro:
* authorize a host (any roles)
* stop hostagent so it no longer pings
* restart bbmaster so it forgets the host
* deauth to see it get stuck
With this change, the deauth completes. Also completed an automerge build.