Improve logging during support bundle retrieval
When pulling a support bundle on a host for the first time, the directory is
searched for existing support bundles. If there is no directory present,
du_ctl will throw an alarming exception.

To improve user experience, this exception has been packed away into a debug
level log message. This message can still be viewed if using the new
"--debug" argument added.