Drop the old method of hostagent restart
In the last release, the hostagent restart (on update of hostagent) was moved
from the post install scripts to the hostagent code itself. Details here:
https://github.com/platform9/pf9-du/commit/019d522bbfa3d6dd422442660b1f1f4862032072

Now that this method is active for all the deployments, it is safe to remove the
old restart method.

Testing done:
http://teamcity.platform9.sys:8111/viewLog.html?buildId=505288&buildTypeId=Pf9project_IntegrationBuildPipeline_016upgradeTests&tab=buildResultsDiv.
The failures are not related to the actual host agent upgrade.

Issues: IAAS-8474, IAAS-8378

Review: https://rbcommons.com/s/platform9/r/8150/