Patch muster deb files as part of hostagent upgrade
We have a release of bad muster deb package scripts already
out to customers. The side effect of this is that users cannot remove the muster
package. This completes blocks host upgrades.

More details can be found in the JIRA below, but it was due to a bug with fpm
which was exposed when we introduced some post upgrade hook/code for muster in
3.3.

The change patches muster's deb prerm and postinstall scripts as part of the
hostagent update. This will ensure that when the host apps are being moved to
the new version, muster will uninstall successfully.

This workaround should go away post 3.4.

Issues: IAAS-8635

Review: https://rbcommons.com/s/platform9/r/8326/

Testing done:
3.2->3.3 upgrade:
http://teamcity.platform9.sys:8111/viewLog.html?buildId=522924&buildTypeId=Pf9project_IntegrationBuildPipeline_016upgradeTests3233&tab=buildResultsDiv

3.3->ToT upgrade: Failures are unrelated to my change but have also manually
verified muster uninstall and upgrades work.
http://teamcity.platform9.sys:8111/viewLog.html?buildId=522882&buildTypeId=Pf9project_IntegrationBuildPipeline_016upgradeTests&tab=buildResultsDiv