PAL-384: Allow passing app_config to .*events.py
In case of VMware deployments, admin can disable one single vCenter
cluster from UI by editting the cluster_name field in customizable
settings. This requires a PUT on the ostackhost role which in turn calls
the on_auth event. Since on_auth event is invoked, none of the
operations related to cleaning up of the disabled cluster are performed.
To correctly perform the cleanup of the disabled clusters only requires
access to the current app_config. This change allows apps to define if
the auth deauth module needs to be invoked with the current app_config.