No certs host installer build for ubuntu
With this change, a new host installer can be built without certs for debian
hosts. Also, necessary plumbing is added to enable the deb installer to work
with resmgr and nginx containers we build. The change also had to pull in the
necessary changes for debian version of the comms agent to be used by the
installer.

This change is heavily inspired by the changes which had been done for the rpm
equivalent of this feature.

Testing: Ran testbed creation locally and the CSPi testbed came up successfully.

Review: https://rbcommons.com/s/platform9/r/8643/

Related reviews: https://rbcommons.com/s/platform9/r/8641,
https://rbcommons.com/s/platform9/r/8642