If vouch is available on the DU, use it to fetch hostagent certs
Added a cert management module to hostagent and the code to call it from
the installer. First it checks for vouch by pulling the version info.
If it's there, it generates a fresh key and csr, and sends it to vouch
for signing. The certs are then put in place, comms is restarted and
a test connection is made through the tunnel to the interal vouch-private
service.

Also includes code to backup the old key/certs, and restore them if
we can't get a connection through comms. We restore the backups,
restart comms and try another test connection.

I've added this code as part of the hostagent package so that when we do
a cert refresh on an existing hostagent installation, we can do it using
the same bbslave/certs module.

Also, added pip to the hostagent virtualenv, updated setuptools (needed this
for cryptography) and installed the various pf9-du modules in a single pip
install command.

testing:
+ Created a new resmgr image with the new hostagent, and ran through
  the installation. Also tested refreshing the certs on an exising host using
  the 'host-certs refresh' tool.
+ decco and cspi integration builds
+ automerge-pf9-main build