Force python extensions to statically link to python
After updating the hostagent build to use a new setuptools version
and pip, I realized that it was building extensions with a dependency
on the shared libpython.so.1.0. This is evidently a difference between
pip and pure setuptools installations. Our custom statically-linked python
installation doesn't have this library, so it would break on some platforms,
like ubuntu16.

Fixed by forcing setup.py's build_ext tool to use a libpath that
only includes our custom, statically linked python installation. With this
it links to libpython.a instead, and the dependency goes away.

testing: automerge integration build, and cspi tests.