Pin dependency versions in hostagent
The latest version of python cryptography requires OPENSSL_1.0.2 which
is not supported in the openssl library we package with hostagent. I think
it worked in some cases because the native openssl was new enough to
hide the problem, but it failed on a decco testbed centos7 host with
OpenSSL 1.0.1e-fips 11 Feb 2013. This change fixes cryptography at an
earlier version that requires only 1.0.1. Also pinned versions of other
dependencies at their current level.

Also moved the configure_comms call in the hostagent installer so the env
files are available when we request certs and test the comms connection.

testing: automerge, cspi, decco builds