Don't push configuration to bbmaster from request thread, only from process_hosts
fix for https://platform9.atlassian.net/browse/IAAS-8990

First add a unit test that reproduces the problem. Unlike the other
unit tests, this one runs the process_hosts loop in the background
of the test with a short poll interval. Then in the foreground, we fire
add_host (PUT) requests at random intervals. Without this change,
it fails every time at a random point in the 0-50 iteration, with the
role version going backwards instead of forward as Paavan described in
the bug.

The change adds a timed wait on an queue to the process_hosts poller thread
in place of the old sleep. This lets the request threads fire the bbone
poller on-demand to move the state machine forward. _process_existing_hosts
uses the _advance_from_transient_state function (formerly
_recover_unexpected_role_state) to advance the state machine to the point
that a bbmaster push can happen.

After that the normal process_hosts processing takes over.

Note that all the unit tests now need a call to bbone.process_hosts since
add_/delete_role no longer does the push.

testing:
+ new unit test that repros the bug
+ automerge test
+ image build
+ smoke
+ upgrade test.