Substitute host_config into event spec after customizable params
fixes: https://platform9.atlassian.net/browse/IAAS-9063

Moved the code that fills in the event_spec host_config to a
point just before the event is called. This ensures that any customizable
params end up in the host_config doc that's passed to the event.

testing: fixed unit test and made sure it failed without this change.
    Also ran vmware test 010