Add consul role watch to resmgr
To remove the need for roles to be baked into the resmgr container,
allow them to be provided through consul. A thread in resmgr watches
the customers/<customerid>/regions/<regionid>/roles key in consul
for changes, and make sure that roles or role parameters that appear
there are loaded. If consul is unavailable, the monitor isn't
instantiated.

Also added a rudimentary ability for resmgr to load auth_events
python modules from a web server.

TODO:
+ add the role config from qbert's init-region
+ remove the resmgr build's dependency on qbert
+ allow host packages to be configured through a role parameter, and
  make the kube role package available on cloudfront. This will allow
  us to remove the nginx dependency on qbert as well
+ similar nova, glance, cinder etc. These need auth events to be made
  available through a url as well.

testing:
+ automerge to make sure nothing's broken in normal DUs.
+ new unit tests with reasonable coverage
+ instantiated a decco testbed and used this functionality to manuallly
  upgrade the qbert role in the database.