Add the ability to retry consul role add if params are missing
With the new consul resmgr role configuration, it's possible for a role
change to be discovered before the associated parameter changes needed
to validate the role. A role can't be saved in the resmgr db unless all
the replaceable params in the role are available. To deal with this,
when validation of a role discovered in consul fails, save it and retry
on the next parameter update. A new parameter may make it possible to
save the role.

Also fixed a problem with config section creation.

testing: cspi build