Increase the timeout for hostagent to intialize to 60 secs
After the hostagent starts, there is a bit of processing that is done before it
has fully initialized and ready to interact with the management plane. The
installer used to wait for 30s to see the existence of some state dir on the
host to indicate that initialization has completed. This is insufficient on slow
hosts (could be slow package manager,etc.).

This change increases the timeout to 60s to give hostagent enough time to
initialize. Having a larger timeout should not lead to side effects.

Issues: PMK-1476

Review: https://rbcommons.com/s/platform9/r/9850/