Fix bbmaster and bbone  builds
THIS IS A HACK! More of a quick fix to get the builds going again.

The python modules used in bbmaster and the bbone integration test are not tied
with a constraints file. Some dependent python modules changed upstream recently
which makes it incompatible to work with older versions of setuptools.

This change updates the version of setuptools inside the virtual env we use to a
newer one. Note that the right change would be to use a constraints file when
installing the python modules in the virtual env. That can be a subsequent fix.

Testing done: http://teamcity.platform9.sys:8111/viewLog.html?buildId=793679