Add resmgr DB migration step to handle kube role url change
Some releases back, in the pf9-kube role, the role json file was made to
use the URL as a replaceable token during the role apply. The value of the role
comes from the a setting specified in the confd directory. This leads to newer
versions of the rpm to be pushed to hosts even when the host is meant to have an
older version of the role.

To address this, the roles in the DB need to be updated to use a fixed version
rpm to download in the URL. Since we already have some releases out with this
problem, the migration script should handle those versions.

Testing: Ran the migration script for resmgr and ensured the DB is updated.

Review: https://rbcommons.com/s/platform9/r/10136/

Issues: CORE-663, PMK-1624